Fix for the Sopwith Camel YASim aircraft (sopwithCamel-YASim) for the FGAddon split.
See the thread at http://thread.gmane.org/gmane.games.flightgear.devel/77465 for
discussions.

After the split of the FGAddon repository from FGData, the sopwithCamel-YASim
aircraft would fail to start, instead showing a "FATAL" error.  The problem was
the use of a relative path to ../Generic/ which no longer exists.  Changing this
to Aircraft/Generic/ in the XML file fixes the path issue.

git-svn-id: http://svn.code.sf.net/p/flightgear/fgaddon/trunk/Aircraft/sopwithCamel@520 9eb9bc84-db9d-4e05-893b-26dd41ac94a6